fix tests * use PSR-4 * consistently use phpunit's testcase as base class * small doc block fixes
