+++ tests/AppInfo/ApplicationTest.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * @author Christoph Wurst <christoph@winzerhof-wurst.at>
+ *
+ * Mail
+ *
+ * This code is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License, version 3,
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ *
+ */
+
+namespace OCA\Mail\Tests\AppInfo;
+
+use OCA\Mail\AppInfo\Application;
+use PHPUnit_Framework_TestCase;
+
+class ApplicationTest extends PHPUnit_Framework_TestCase {
+
+	public function testConstrucor() {
+		// Not really a test – it's just about code coverage
+		new Application();
+	}
+
+}
