+++ tests/HordeTranslationHandlerTest.php
