+++ tests/Service/Autocompletion/AddressCollectorTest.php
