+++ tests/Service/LoggerTest.php
@@ -1,0 +1,64 @@
+<?php
+
+/**
+ * @author Christoph Wurst <christoph@winzerhof-wurst.at>
+ * @author Thomas Müller <thomas.mueller@tmit.eu>
+ *
+ * Mail
+ *
+ * This code is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License, version 3,
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ *
+ */
+
+namespace OCA\Mail\Tests\Service;
+
+use Exception;
+use OCA\Mail\Service\Logger;
+use PHPUnit_Framework_TestCase;
+
+class LoggerTest extends PHPUnit_Framework_TestCase {
+
+	/**
+	 * @dataProvider providesLoggerMethods
+	 * @param $method
+	 */
+	public function testLoggerMethod($method, $param = '1') {
+
+		$baseLogger = $this->getMockBuilder('\OCP\ILogger')->getMock();
+		$baseLogger->expects($this->once())
+			->method($method)
+			->with(
+				$this->equalTo($param), $this->equalTo([
+					'app' => 'mail',
+					'key' => 'value',
+		]));
+
+		$logger = new Logger('mail', $baseLogger);
+		$logger->$method($param, ['key' => 'value']);
+	}
+
+	public function providesLoggerMethods() {
+		return [
+			['alert'],
+			['warning'],
+			['emergency'],
+			['critical'],
+			['error'],
+			['notice'],
+			['info'],
+			['debug'],
+			['logException', new Exception()],
+		];
+	}
+
+}
